implementation de la couche controller, entitie, service et repository de l'entité facture

+++ backend/serverFacturation/src/main/java/sever/com/serverFacturation/dtos/UserDto.java
@@ -24,7 +24,7 @@ public class UserDto {
     private String prenom;
     private String email;
     private String passwd;
-    private BigDecimal solde;
+    private Double solde;
     private String telephone;
     private LocalDate dateNaiss;
     private TypeRoles roles;
